python_backend/requirements.txt
pythonocc-core>=7.7.0
numpy>=1.24.0
